community/members: fix overflow in ProfileCard
All ProfileCards are now the same width and height. This is not possible
with flexbox, so it has been changed to use grid.
The Makefile has also been updated to use `npx gatsby ...` instead of
`gatsby ...`. This removes the need for installing gatsby-cli globally.
Additionally, there was a stray meshmate_of_the_year svg which was not
being used anywhere, and has been removed.

Signed-off-by: Pranjal Kole <pranjal.kole7@gmail.com>